==> 1.Combining_Raw_Data.ipynb <==
# Import libraries

import pandas as pd

# Read in city csvs

first = pd.read_csv('raw_data\first.csv')
second = pd.read_csv('raw_data\atlanta.csv')
third = pd.read_csv('raw_data\austin.csv')
fourth = pd.read_csv('raw_data\boston.csv')
fifth = pd.read_csv('raw_data\chicago.csv')
sixth = pd.read_csv('raw_data\denver.csv')
seventh = pd.read_csv('raw_data\kansas city.csv')
eighth = pd.read_csv('raw_data\los angeles.csv')
ninth = pd.read_csv('raw_data\minneapolis.csv')
tenth = pd.read_csv('raw_data\portland.csv')
eleventh = pd.read_csv('raw_data\san francisco.csv')

# Combine dataframes

raw_df = pd.concat([first,second,third,fourth,fifth,sixth,seventh,eighth,ninth,tenth,eleventh])

raw_df.shape

# Export csv

raw_df.to_csv('raw_df.csv')
