==> 2.Transforming_To_Splits.ipynb <==
# Import Libraries

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import warnings
warnings.filterwarnings('ignore')

# Read in Raw Data

raw_df = pd.read_csv('raw_df.csv', index_col=[0])

raw_df.columns

# Create Target Variable

#creates boolean column that shows if there was any precipitation at all
raw_df['did_it_rain_today'] = raw_df['precip'].apply(lambda x: 1 if x>0 else 0)
#shifts the did it rain column down to show that it will rain tomorrow
raw_df['target'] = raw_df.groupby('name')['did_it_rain_today'].shift()
#remomving values where we dont have the previous day
df = raw_df[raw_df['target'].notnull()]

df.drop('did_it_rain_today',inplace=True,axis=1)

df['datetime'] = pd.to_datetime(df['datetime'])

df.info()

We will split the data into a train, validation, and test set based on the time. 

training_data = df[(df['datetime'] <= '2021-05-28')]
valid_data = df[(df['datetime'] > '2021-05-28') & (df['datetime'] <= '2021-10-26')]
test_data = df[(df['datetime'] > '2021-10-26')]

training_data.to_csv('training_data.csv')
valid_data.to_csv('valid_data.csv')
test_data.to_csv('test_data.csv')
