==> 4.Model Selection .ipynb <==
# Import libraries

import pandas as pd
import numpy as np
from datetime import datetime as dt
from sklearn.preprocessing import StandardScaler
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import f1_score
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import BernoulliNB
from sklearn.ensemble import VotingClassifier
import pickle as pkl
from sklearn.model_selection import GridSearchCV

# Prepare the training, validation, and test data

raw_training = pd.read_csv('training_data.csv',index_col=[0])
raw_valid = pd.read_csv('valid_data.csv',index_col=[0])
raw_test = pd.read_csv('test_data.csv',index_col=[0])

List_of_changes for pipeline <br>
1. Add temp more than feels like <br>
2. Drop preciptype, precipprob, severerisk, and description, icon, stations,tempmin,tempmax,temp,feelslikemax,feelslikemin,feelslike <br>
3. Add if precipcover is over 75 <br>
4.Scale features <br> 
5. Most important: cloudcover, humidity <br>
6. Convert sunset and sunrise
7. Create number of conditions column

# Data Cleaning

raw_training.drop(['preciptype','precipprob','severerisk','description','icon','stations','tempmin','tempmax','temp','feelslikemax','feelslikemin','feelslike'],axis=1, inplace=True)
raw_valid.drop(['preciptype','precipprob','severerisk','description','icon','stations','tempmin','tempmax','temp','feelslikemax','feelslikemin','feelslike'],axis=1, inplace=True)
raw_test.drop(['preciptype','precipprob','severerisk','description','icon','stations','tempmin','tempmax','temp','feelslikemax','feelslikemin','feelslike'],axis=1, inplace=True)

# Feature Engineering

#this feature will represent the difference between what it feels like and the actual temperature
raw_training['feels_more_than_actual'] = raw_training['feelslike'] - raw_training['temp']
#this feature will reprensent the range between the maximum and minimum temperature for the day
raw_training['range_of_temp'] = raw_training['tempmax'] - raw_training['tempmin']
raw_valid['feels_more_than_actual'] = raw_valid['feelslike'] - raw_valid['temp']
raw_valid['range_of_temp'] = raw_valid['tempmax'] - raw_valid['tempmin']
raw_test['feels_more_than_actual'] = raw_test['feelslike'] - raw_test['temp']
raw_test['range_of_temp'] = raw_test['tempmax'] - raw_test['tempmin']

#this feature is a boolean that determines if the percentage of precipation is very high
raw_training['mostly_raining'] = raw_training['precipcover'].apply(lambda x: 1 if x>75 else 0)
raw_valid['mostly_raining'] = raw_valid['precipcover'].apply(lambda x: 1 if x>75 else 0)
raw_test['mostly_raining'] = raw_test['precipcover'].apply(lambda x: 1 if x>75 else 0)

raw_training.drop(['precipcover'],axis=1,inplace=True)
raw_valid.drop(['precipcover'],axis=1, inplace=True)
raw_test.drop(['precipcover'],axis=1,inplace=True)

#changes sunrise to the correct timing format
raw_training['sunrise'] = raw_training['sunrise'].apply(lambda x: x[-5:])
raw_training['sunrise'] = raw_training['sunrise'].apply(lambda x: '0'+x if len(x)==4 else x)
raw_training['sunrise'] = raw_training['sunrise'].apply(lambda x: str(int(x[:2])-24)+x[-3:] if int(x[:2]) >= 24 else x)
raw_training['sunrise'] = pd.to_datetime(raw_training['sunrise']).dt.strftime('%H:%M')

raw_valid['sunrise'] = raw_valid['sunrise'].apply(lambda x: x[-5:])
raw_valid['sunrise'] = raw_valid['sunrise'].apply(lambda x: '0'+x if len(x)==4 else x)
raw_valid['sunrise'] = raw_valid['sunrise'].apply(lambda x: str(int(x[:2])-24)+x[-3:] if int(x[:2]) >= 24 else x)
raw_valid['sunrise'] = pd.to_datetime(raw_valid['sunrise']).dt.strftime('%H:%M')

raw_test['sunrise'] = raw_test['sunrise'].apply(lambda x: x[-5:])
raw_test['sunrise'] = raw_test['sunrise'].apply(lambda x: '0'+x if len(x)==4 else x)
raw_test['sunrise'] = raw_test['sunrise'].apply(lambda x: str(int(x[:2])-24)+x[-3:] if int(x[:2]) >= 24 else x)
raw_test['sunrise'] = pd.to_datetime(raw_test['sunrise']).dt.strftime('%H:%M')

#changes sunset to the correct timing format
raw_training['sunset'] = raw_training['sunset'].apply(lambda x: x[-5:])
raw_training['sunset'] = raw_training['sunset'].apply(lambda x: '0'+x if len(x)==4 else x)
raw_training['sunset'] = raw_training['sunset'].apply(lambda x: str(int(x[:2])-24)+x[-3:] if int(x[:2]) >= 24 else x)
raw_training['sunset'] = pd.to_datetime(raw_training['sunset']).dt.strftime('%H:%M')

raw_valid['sunset'] = raw_valid['sunset'].apply(lambda x: x[-5:])
raw_valid['sunset'] = raw_valid['sunset'].apply(lambda x: '0'+x if len(x)==4 else x)
raw_valid['sunset'] = raw_valid['sunset'].apply(lambda x: str(int(x[:2])-24)+x[-3:] if int(x[:2]) >= 24 else x)
raw_valid['sunset'] = pd.to_datetime(raw_valid['sunset']).dt.strftime('%H:%M')

raw_test['sunset'] = raw_test['sunset'].apply(lambda x: x[-5:])
raw_test['sunset'] = raw_test['sunset'].apply(lambda x: '0'+x if len(x)==4 else x)
raw_test['sunset'] = raw_test['sunset'].apply(lambda x: str(int(x[:2])-24)+x[-3:] if int(x[:2]) >= 24 else x)
raw_test['sunset'] = pd.to_datetime(raw_test['sunset']).dt.strftime('%H:%M')

X_train['sunrise_minutes'] = X_train['sunrise'].apply(lambda x: int(x[:2])*60+int(x[-2:]))
X_valid['sunrise_minutes'] = X_valid['sunrise'].apply(lambda x: int(x[:2])*60+int(x[-2:]))
X_test['sunrise_minutes'] = X_test['sunrise'].apply(lambda x: int(x[:2])*60+int(x[-2:]))

X_train['sunset_minutes'] = X_train['sunset'].apply(lambda x: int(x[:2])*60+int(x[-2:]))
X_valid['sunset_minutes'] = X_valid['sunset'].apply(lambda x: int(x[:2])*60+int(x[-2:]))
X_test['sunset_minutes'] = X_test['sunset'].apply(lambda x: int(x[:2])*60+int(x[-2:]))

#creates an int of how many conditions for the day are listed
raw_training['conditions'] = raw_training['conditions'].apply(lambda x: len(x.split(',')))
raw_valid['conditions'] = raw_valid['conditions'].apply(lambda x: len(x.split(',')))
raw_test['conditions'] = raw_test['conditions'].apply(lambda x: len(x.split(',')))

X_train.drop(['name','datetime'],inplace=True,axis=1)
X_valid.drop(['name','datetime'],inplace=True,axis=1)
X_test.drop(['name','datetime'],inplace=True,axis=1)

raw_training.reset_index(drop=True,inplace=True)
raw_valid.reset_index(drop=True,inplace=True)
raw_test.reset_index(drop=True,inplace=True)


y_train = raw_training['target']
X_train = raw_training.drop(['target'],inplace=True,axis=1)
y_valid = raw_valid['target']
X_valid = raw_valid.drop(['target'],inplace=True,axis=1)
y_test = raw_test['target']
X_test = raw_test.drop(['target'],inplace=True,axis=1)


X_train.drop(['sunset','sunrise'],inplace=True,axis=1)
X_valid.drop(['sunset','sunrise'],inplace=True,axis=1)
X_test.drop(['sunset','sunrise'],inplace=True,axis=1)

# Preprocessing Data

scaler = StandardScaler()
imputer = SimpleImputer()

#imputer missing values
X_train = imputer.fit_transform(X_train)
X_valid = imputer.transform(X_valid)
X_test = imputer.transform(X_test)

#scale data
X_train = scaler.fit_transform(X_train)
X_valid = scaler.transform(X_valid)
X_test = scaler.transform(X_test)

y_train = y_train.astype('category')
#df['target'] = df['target'].astype('object')
y_valid = y_valid.astype('category')

Adding Polynomial features

poly = PolynomialFeatures(degree=2)
X_train_poly = poly.fit_transform(X_train)
X_valid_poly = poly.fit_transform(X_valid)
X_test_poly = poly.fit_transform(X_test)

# Build Baseline Models

linear_model = LogisticRegression()
linear_model_poly = LogisticRegression()

linear_model.fit(X_train,y_train)

f1_score(y_valid,linear_model.predict(X_valid))
f1_score(y_valid,linear_model_poly.predict(X_valid_poly))

Lets get baseline scores of each model

kneighbors = KNeighborsClassifier()
kneighbors.fit(X_train,y_train)
f1_score(y_valid,kneighbors.predict(X_valid))

k_range = list(range(20, 40))
param_grid = dict(n_neighbors=k_range)

grid = GridSearchCV(kneighbors, param_grid, cv=10)
grid.fit(X_train,y_train)

grid.best_params_

grid.best_score_

kneighbors = KNeighborsClassifier(n_neighbors=34)
kneighbors.fit(X_train,y_train)
f1_score(y_valid,kneighbors.predict(X_valid))

best_score = 0
best_param = 0
for i in range(1,20):
    kneighbors = KNeighborsClassifier(n_neighbors=i)
    kneighbors.fit(X_train,y_train)
    current_score = f1_score(y_valid,kneighbors.predict(X_valid))
    if current_score > best_score:
        best_score = current_score
        best_param = i
        

best_param

Best score on validation set with KKN where n_neighbors is 9

best_score

Lets move on to Decision Tree

decision_tree = DecisionTreeClassifier(criterion='gini', max_depth=3, random_state=0)
decision_tree.fit(X_train,y_train)
f1_score(y_valid,decision_tree.predict(X_valid))

tree_param = {'criterion': ['entropy','gini'], 'max_depth': [2,3,4,5,6,7,8,9],'min_samples_leaf':[1,2,3,4,5]}

decision_tree = DecisionTreeClassifier()

grid_search_decision_tree = GridSearchCV(decision_tree,tree_param, cv=5)

grid_search_decision_tree.fit(X_train,y_train)

grid_search_decision_tree.best_params_

random_forest = RandomForestClassifier()

random_param = {'criterion': ['entropy','gini'], 'max_depth': [2,3,4,5,6,7,8,9],'min_samples_leaf':[1,2,3,4,5],'n_estimators':[50,100,150,200]}

grid_search_random = GridSearchCV(random_forest,random_param,cv=5)

grid_search_random.fit(X_train,y_train)

grid_search_random.best_params_

grid_search_random.best_score_

f1_score(y_valid,grid_search_random.predict(X_valid))

random_forest = RandomForestClassifier(criterion='gini',max_depth=4,min_samples_leaf=3,n_estimators=100)
random_forest.fit(X_train_poly,y_train)
f1_score(y_valid,random_forest.predict(X_valid_poly))

We will not move forward with the polynomial features because their incremental improvement on f1 is not worth the added computational complexity.

naive_bayes = BernoulliNB()
naive_bayes.fit(X_train,y_train)

f1_score(y_valid,naive_bayes.predict(X_valid))

# Ensemble Method

voting_classifier = VotingClassifier(estimators=[('nb',naive_bayes),('rf',decision_tree),('lr',linear_model)],weights=[1,.75,.75])

voting_classifier.fit(X_train,y_train)

f1_score(y_valid,voting_classifier.predict(X_valid))

f1_score(y_test,voting_classifier.predict(X_test))

with open('final_model.pickle','wb') as modelfile:
    pkl.dump(voting_classifier,modelfile)
